--- a/snappy/snappy.go
+++ b/snappy/snappy.go
@@ -1,11 +1,12 @@
 package snappy
 
 import (
-	"bytes"
 	"encoding/binary"
+	"io"
+	"sync"
 
 	"github.com/golang/snappy"
-	"github.com/segmentio/kafka-go"
+	kafka "github.com/segmentio/kafka-go"
 )
 
 func init() {
@@ -23,26 +24,73 @@ func NewCompressionCodec() CompressionCodec {
 }
 
 // Code implements the kafka.CompressionCodec interface.
-func (c CompressionCodec) Code() int8 {
+func (CompressionCodec) Code() int8 {
 	return Code
 }
 
 // Encode implements the kafka.CompressionCodec interface.
-func (c CompressionCodec) Encode(src []byte) ([]byte, error) {
+func (CompressionCodec) Encode(src []byte) ([]byte, error) {
 	// NOTE : passing a nil dst means snappy will allocate it.
 	return snappy.Encode(nil, src), nil
 }
 
 // Decode implements the kafka.CompressionCodec interface.
-func (c CompressionCodec) Decode(src []byte) ([]byte, error) {
+func (CompressionCodec) Decode(src []byte) ([]byte, error) {
 	return decode(src)
 }
 
-var xerialHeader = []byte{130, 83, 78, 65, 80, 80, 89, 0}
+// NewReader implements the kafka.CompressionCodec interface.
+func (CompressionCodec) NewReader(r io.Reader) io.ReadCloser {
+	x := readerPool.Get().(*xerialReader)
+	x.Reset(r)
+	return &reader{x}
+}
+
+// NewWriter implements the kafka.CompressionCodec interface.
+func (CompressionCodec) NewWriter(w io.Writer) io.WriteCloser {
+	x := writerPool.Get().(*xerialWriter)
+	x.Reset(w)
+	return &writer{x}
+}
+
+type reader struct{ *xerialReader }
+
+func (r *reader) Close() (err error) {
+	if x := r.xerialReader; x != nil {
+		r.xerialReader = nil
+		x.Reset(nil)
+		readerPool.Put(x)
+	}
+	return
+}
+
+type writer struct{ *xerialWriter }
+
+func (w *writer) Close() (err error) {
+	if x := w.xerialWriter; x != nil {
+		w.xerialWriter = nil
+		err = x.Flush()
+		x.Reset(nil)
+		writerPool.Put(x)
+	}
+	return
+}
+
+var readerPool = sync.Pool{
+	New: func() interface{} {
+		return &xerialReader{decode: snappy.Decode}
+	},
+}
+
+var writerPool = sync.Pool{
+	New: func() interface{} {
+		return &xerialWriter{encode: snappy.Encode}
+	},
+}
 
 // From github.com/eapache/go-xerial-snappy
 func decode(src []byte) ([]byte, error) {
-	if !bytes.Equal(src[:8], xerialHeader) {
+	if !isXerialHeader(src) {
 		return snappy.Decode(nil, src)
 	}
 
@@ -53,6 +101,7 @@ func decode(src []byte) ([]byte, error) {
 		chunk []byte
 		err   error
 	)
+
 	for pos < max {
 		size := binary.BigEndian.Uint32(src[pos : pos+4])
 		pos += 4
@@ -64,5 +113,6 @@ func decode(src []byte) ([]byte, error) {
 		pos += size
 		dst = append(dst, chunk...)
 	}
+
 	return dst, nil
 }