--- a/zstd/zstd.go
+++ b/zstd/zstd.go
@@ -42,6 +42,11 @@ func (c CompressionCodec) Code() int8 {
 	return Code
 }
 
+// Name implements the kafka.CompressionCodec interface.
+func (c CompressionCodec) Name() string {
+	return "zstd"
+}
+
 // NewReader implements the kafka.CompressionCodec interface.
 func (c CompressionCodec) NewReader(r io.Reader) io.ReadCloser {
 	return &reader{