This PR modifies the compression codec API to employ `io.Reader` and `io.Writer` instead of working with byte slices. The upside is the amount of memory that gets allocated internally is greatly reduced, and we pave the way for better internal abstractions to build more efficient programs with kafka-go.

This is a breaking change in the API, but I think it's OK since these packages are intended to work with kafka-go, and programs shouldn't use these directly (maybe we'll bump to 0.3.0 on the next release?).

Here are comparisons of the compression benchmarks taken before and after:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkCompression/GZIP1024        48754         57090         +17.10%
BenchmarkCompression/GZIP4096        116650        110110        -5.61%
BenchmarkCompression/GZIP8192        204592        161450        -21.09%
BenchmarkCompression/GZIP16384       388270        290743        -25.12%
BenchmarkCompression/Snappy1024      533           509           -4.50%
BenchmarkCompression/Snappy4096      1287          1055          -18.03%
BenchmarkCompression/Snappy8192      2361          1732          -26.64%
BenchmarkCompression/Snappy16384     4429          3427          -22.62%
BenchmarkCompression/LZ41024         1033620       2991          -99.71%
BenchmarkCompression/LZ44096         994821        13117         -98.68%
BenchmarkCompression/LZ48192         1063436       28193         -97.35%
BenchmarkCompression/LZ416384        1059942       67983         -93.59%
BenchmarkCompression/zstd1024        11835         12423         +4.97%
BenchmarkCompression/zstd4096        23943         27353         +14.24%
BenchmarkCompression/zstd8192        39247         38990         -0.65%
BenchmarkCompression/zstd16384       70687         69344         -1.90%

benchmark                            old MB/s       new MB/s     speedup
BenchmarkCompression/GZIP1024        16.63          14.17        0.85x
BenchmarkCompression/GZIP4096        26.58          28.19        1.06x
BenchmarkCompression/GZIP8192        30.13          38.20        1.27x
BenchmarkCompression/GZIP16384       31.66          42.27        1.34x
BenchmarkCompression/Snappy1024      1927.83        2057.82      1.07x
BenchmarkCompression/Snappy4096      3185.87        3904.71      1.23x
BenchmarkCompression/Snappy8192      3471.45        4743.12      1.37x
BenchmarkCompression/Snappy16384     3700.07        4788.41      1.29x
BenchmarkCompression/LZ41024         1.01           348.62       345.17x
BenchmarkCompression/LZ44096         4.14           313.71       75.78x
BenchmarkCompression/LZ48192         7.72           291.23       37.72x
BenchmarkCompression/LZ416384        15.48          241.28       15.59x
BenchmarkCompression/zstd1024        68.44          65.20        0.95x
BenchmarkCompression/zstd4096        129.22         113.22       0.88x
BenchmarkCompression/zstd8192        156.60         157.73       1.01x
BenchmarkCompression/zstd16384       173.44         176.77       1.02x

benchmark                            old allocs     new allocs     delta
BenchmarkCompression/GZIP1024        5              4              -20.00%
BenchmarkCompression/GZIP4096        7              4              -42.86%
BenchmarkCompression/GZIP8192        8              4              -50.00%
BenchmarkCompression/GZIP16384       9              4              -55.56%
BenchmarkCompression/Snappy1024      2              2              +0.00%
BenchmarkCompression/Snappy4096      2              2              +0.00%
BenchmarkCompression/Snappy8192      2              2              +0.00%
BenchmarkCompression/Snappy16384     2              2              +0.00%
BenchmarkCompression/LZ41024         18             2              -88.89%
BenchmarkCompression/LZ44096         20             2              -90.00%
BenchmarkCompression/LZ48192         21             2              -90.48%
BenchmarkCompression/LZ416384        22             2              -90.91%
BenchmarkCompression/zstd1024        4              2              -50.00%
BenchmarkCompression/zstd4096        4              2              -50.00%
BenchmarkCompression/zstd8192        4              2              -50.00%
BenchmarkCompression/zstd16384       4              2              -50.00%

benchmark                            old bytes     new bytes     delta
BenchmarkCompression/GZIP1024        4079          110           -97.30%
BenchmarkCompression/GZIP4096        23417         153           -99.35%
BenchmarkCompression/GZIP8192        49863         154           -99.69%
BenchmarkCompression/GZIP16384       101673        243           -99.76%
BenchmarkCompression/Snappy1024      2304          16            -99.31%
BenchmarkCompression/Snappy4096      8960          16            -99.82%
BenchmarkCompression/Snappy8192      17920         16            -99.91%
BenchmarkCompression/Snappy16384     36864         16            -99.96%
BenchmarkCompression/LZ41024         17316671      68            -100.00%
BenchmarkCompression/LZ44096         17341247      278           -100.00%
BenchmarkCompression/LZ48192         17374015      541           -100.00%
BenchmarkCompression/LZ416384        17439551      1331          -99.99%
BenchmarkCompression/zstd1024        2224          65            -97.08%
BenchmarkCompression/zstd4096        9008          66            -99.27%
BenchmarkCompression/zstd8192        17712         66            -99.63%
BenchmarkCompression/zstd16384       34864         71            -99.80%
```

One aspect of this change that I would like to open the conversation on is framing with snappy. We used to support xerial framing on the reader side only, here I implemented framing on the writer side as well, which means the batches produces to kafka will be slightly different than what they are today. Since we were compatible on the reader side already this shouldn't be an issue, however I'm thinking we should probably make it an opt-in feature and still default to the historical behavior by default.

Another aspect is the framing mechanism used by the `github.com/golang/snappy` package is different than the one typically used in kafka programs (xerial), so the snappy-encoded data that we produce cannot be decoded with a `*snappy.Reader`. I wonder if we should offer the ability to support both framing formats.

Let me know what you think!