@Pryz 2019-07-09
Maybe a good opportunity to move to https://github.com/klauspost/compress/tree/master/zstd ?