@achille-roussel 2019-07-09
We had https://github.com/segmentio/kafka-go/issues/303 opened recently for this, let's give it a bit of time?