@Pryz 2019-07-09
> Another aspect is the framing mechanism used by the github.com/golang/snappy package is different than the one typically used in kafka programs (xerial), so the snappy-encoded data that we produce cannot be decoded with a *snappy.Reader. I wonder if we should offer the ability to support both framing formats.

I believe the xerial framing only matter if we let Kafka doing the compression. When using `compression.type` this may not matter. We should do some testing tho.