@achille-roussel 2019-07-09
Cool, probably not worth investigating how to get compatibility with the `github.com/golang/snappy` reader and writer then 👍