@stevevls 2019-07-12
Re: the xerial framing, I would take a look at the java client, librdkafka, and maybe even sarama to see what they are able to produce/consume.  I think we should stick to whatever is supported by the Kafka ecosystem at large.  We don't want to run into situations where folks are deploying kafka-go alongside other kafka connectors and find that they can't talk to one another! 😅