@stevevls 2019-07-12
Re: the breaking change, I think it's totally fine.  I don't think there's a strong case for people to be writing their own codecs since any compression formats require broker support and a predefined constant.