@achille-roussel 2019-07-12
@stevevls good point on comparing with other libraries.

**kafka-python** makes it optional, but ON by default:
https://github.com/dpkp/kafka-python/blob/master/kafka/codec.py#L101-L124

**Sarama** forces Xenial framing: https://github.com/Shopify/sarama/blob/94536b3e82d393e4b5dfa36475b40fc668e9486f/compress.go#L55

**librdkafka** seem to never output with framing enabled: https://github.com/edenhill/librdkafka/blob/bd74dc31ba6edf431749b4ad5a00aa4fc78882fc/src/snappy.c#L1408-L1484
but supports decoding with the xerial framing: https://github.com/edenhill/librdkafka/blob/bd74dc31ba6edf431749b4ad5a00aa4fc78882fc/src/snappy.c#L1583-L1694

So I guess we're pretty much free to pick whatever approach works for us 🤷‍♂