@stevevls 2019-07-12
Just to make things more interesting, the java client uses Xerial.  That's probably how all this craziness started:

https://github.com/apache/kafka/blob/741cb761c5239297029a446518c332f6c4ed08f6/clients/src/main/java/org/apache/kafka/common/record/CompressionType.java#L210-L214