@stevevls 2019-07-12
Looking at the above, it seems to me that we want to support decoding both framing formats for sure.

As for encoding...it seems like the best approach for interoperability would be to encode with xerial framing.  However, as you point out, that would not be strictly backwards compatible.  I think one could make the argument that it *shouldn't* cause operational issues given that all other known clients would be able to read the xerial-framed records.

I would say that if we're more concerned with *strict* b/w compatibility, let's default to unframed with an option to enable xerial.  Otherwise, unless there's a big advantage to supporting both, I'd suggest we make our lives simple by just supporting xerial.  Admittedly, I don't know enough about snappy to know whether one is better than the other. 😄