@achille-roussel 2019-07-13
Defaulting to "unframed" means we have to buffer the whole compressed batch in the intermediary compression buffer, so we default to the unoptimized approach.

Defaulting to "framed" means we retain the historical behavior.

My understanding is Kafka will always return a framed snappy compressed batch, even if it was produced without framing (with a single frame, which is why we had to support it). Although we can do more testing to validate this behavior.