@achille-roussel 2019-07-13
Alright, I did some cleanup of the code, made framing configurable and the default in the snappy codec, and improved usability of the benchmarks, here's what they look like now:
```
$ go test -v -run _ -bench BenchmarkCompression -benchmem
goos: darwin
goarch: amd64
pkg: github.com/segmentio/kafka-go
BenchmarkCompression/none/compress         	   20000	     72263 ns/op	26852.58 MB/s	      16 B/op	       1 allocs/op
BenchmarkCompression/none/decompress       	   50000	     35302 ns/op	54967.20 MB/s	      16 B/op	       1 allocs/op
BenchmarkCompression/gzip/compress         	     100	  17589721 ns/op	   6.77 MB/s	    8156 B/op	       1 allocs/op
BenchmarkCompression/gzip/decompress       	     500	   2464446 ns/op	 787.39 MB/s	    7631 B/op	      55 allocs/op
BenchmarkCompression/snappy/compress       	    2000	   1146307 ns/op	 230.73 MB/s	      45 B/op	       1 allocs/op
BenchmarkCompression/snappy/decompress     	    2000	    589614 ns/op	3291.08 MB/s	      40 B/op	       1 allocs/op
BenchmarkCompression/lz4/compress          	     500	   2695989 ns/op	  75.24 MB/s	   17850 B/op	       1 allocs/op
BenchmarkCompression/lz4/decompress        	    2000	   1160765 ns/op	1671.72 MB/s	    4206 B/op	       1 allocs/op
BenchmarkCompression/zstd/compress         	     300	   4867299 ns/op	  24.30 MB/s	   13222 B/op	       1 allocs/op
BenchmarkCompression/zstd/decompress       	    2000	    830720 ns/op	2335.89 MB/s	    2624 B/op	       1 allocs/op
input => 1.85 MB
  none:		0.00%
  gzip:		93.86%
  snappy:	86.37%
  lz4:		89.55%
  zstd:		93.90%

PASS
ok  	github.com/segmentio/kafka-go	19.124s
```